requirements.txt
requests
beautifulsoup4
pandas
openai
python-dotenv
tiktoken
